backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
passlib[bcrypt]
pyjwt
pillow
